requirements.txt
requests==2.12.4
gpiozero==1.4.1
word2number==1.1
six==1.10.0
newsapi==0.1.1
twilio==6.10.3
omxplayer_wrapper==0.2.5
PyDictionary==1.5.2
google_api_python_client==1.6.6
paho_mqtt==1.3.1
